requirements.txt
fastf1
Flask
matplotlib
numpy
plotly
pandas
Flask-Mail
python-dotenv
Flask-Cors
